Added modal dialog on teams
