+++ src/components/CompetitionPools/CompetitionPools.scss
@@ -139,6 +139,15 @@
               font-size: 1rem;
             }
 
+            .vb-team-link {
+              margin-left: 0.375rem;
+              font-size: 0.875rem;
+              opacity: 0.85;
+              svg {
+                transform: translateY(-0.0625rem);
+              }
+            }
+
             .strikethrough {
               text-decoration: line-through;
             }
@@ -189,4 +198,8 @@
   .vb-championship {
     color: white;
   }
+}
+
+.ant-modal-wrap .ant-modal-content {
+  padding: 0;
 }
