+++ src/components/MatchSheetLink/MatchSheetLink.scss
@@ -1,0 +1,5 @@
+.vb-match-sheet-link {
+  margin-left: 0.375rem;
+  // font-size: 0.625rem;
+  opacity: 0.85;
+}
